Baekjoon 24264 C++

+++ Baekjoon/24264 - 알고리즘 수업 - 알고리즘의 수행 시간 3/24264.cpp
@@ -2,7 +2,7 @@
 using namespace std;
 
 int main() {
-    int n;
+    long long int n;
     
     cin >> n;
     cout << n*n << "\n2";
